backend/requirements.txt
httpx>=0.27,<1.0
pydantic>=2.5,<3.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
respx>=0.21,<1.0
